+++ base/exports.jl
@@ -586,7 +586,6 @@ export
     linreg,
     logabsdet,
     logdet,
-    logm,
     lu,
     lufact!,
     lufact,