+++ test/linalg/triangular.jl
@@ -174,9 +174,9 @@ for elty1 in (Float32, Float64, BigFloat, Complex64, Complex128, Complex{BigFloa
             @test B == viewA1.'
         end
 
-        #exp/logm
+        #exp/log
         if (elty1 == Float64 || elty1 == Complex128) && (t1 == UpperTriangular || t1 == LowerTriangular)
-            @test exp(full(logm(A1))) ≈ full(A1)
+            @test exp(full(log(A1))) ≈ full(A1)
         end
 
         # scale